[modified] fix code for reading
